bugs fixed

+++ app/src/main/java/ir/ghararemaghzha/game/adapters/ChatAdapter.java
@@ -47,7 +47,7 @@ public class ChatAdapter extends RealmRecyclerViewAdapter<MessageModel, Recycler
 
     @Override
     public long getItemId(int position) {
-        return (long)getItem(position).getMessageId();
+        return getItem(position).getMessageId();
     }
 
     @Override
